+++ frontend/src/components/UseMySeriesPanel.test.tsx
@@ -19,7 +19,6 @@ function makeState(overrides: Partial<ControlsState> = {}): ControlsState {
     genresSelected: [],
     keywordsSelected: [],
     trendingWindow: 'week',
-    minSourceRating: '',
     minTmdbRating: '',
     minVoteCount: '',
     minVoteCountTouched: false,